Make some changes regarding recommendation: * Add findTopByTotalTimesSelected method into ApplicationRepository.java class which also has a nice SQL query (this method is used for TotalTimesSelectedStrategy) * Make slight changes to Recommender.java, Recommendation.java and TotalTimesSelectedStrategy.java class. * Add TotalTimesSelectedStrategyTest.java class that tests TotalTimesSelectedStrategy.java class.

+++ microservices/hiring-procedure/src/main/java/nl/tudelft/sem/hiring/procedure/recommendation/TotalTimesSelectedStrategy.java
@@ -14,7 +14,7 @@ public class TotalTimesSelectedStrategy implements RecommendationStrategy {
     /**
      * Instantiates a new TotalTimesSelectedStrategy object.
      *
-     * @param repo the application repository
+     * @param repo the TA application repository
      */
     public TotalTimesSelectedStrategy(ApplicationRepository repo) {
         this.repo = repo;
@@ -22,19 +22,19 @@ public class TotalTimesSelectedStrategy implements RecommendationStrategy {
 
     /**
      * Recommends at most the specified number of candidate TAs who have applied
-     *   for the specified course.
+     *   for a specified course.
      *   It uses the strategy of the max total times selected for any course.
      *
      * @param courseId      the id of the course
-     * @param number        the maximum number of recommendations for the course
-     * @return the list of recommendations for candidate TAs based on the number of times selected.
-     *         The size of the list is at most 'number'
+     * @param number        the maximum number of recommendations to return
+     * @return the list of recommendations for candidate TAs based on the number of
+     *         times selected for a TA position. The size of the list is at most 'number'.
      */
     public List<Recommendation> recommend(long courseId, int number) {
         return this.repo.findTopByTotalTimesSelected(courseId)
                 .stream()
                 .limit(number)
-                .map(a -> new Recommendation((Long) a[0], (Double) a[1]))
+                .map(a -> new Recommendation((Long) a[0], ((Long) a[1]).doubleValue()))
                 .collect(Collectors.toList());
     }
 }
